+++ Flow/src/utopia/flow/view/mutable/caching/ListenableResettableLazy.scala
@@ -92,7 +92,7 @@ class ListenableResettableLazy[A](generator: => A) extends ResettableLazy[A] wit
 		resetListeners = resetListeners.filterNot { _ == listener }
 	}
 	
-	override def map[B](f: A => B) =
+	override protected def mapToListenable[B](f: A => B) =
 	{
 		val newLazy = ListenableResettableLazy { f(value) }
 		addResetListener(LazyResetListener.onAnyReset { newLazy.reset() })